Fix Redis PHP 8+ type error: cast database numbers to int
Redis::select() in PHP 8+ requires an integer argument, but was receiving
strings from getDatabases() and  parameters.

Fixed:
- getDatabases() now returns int values: (int)str_replace('db', '', $key)
- flushDB() now casts parameter to int: select((int)$db)

Resolves: Fatal error in db_redis.php on line 117

+++ .devilbox/www/include/lib/container/Redis.php
@@ -78,7 +78,7 @@ class Redis extends BaseClass implements BaseInterface
 	public function flushDB($db)
 	{
 		if ($this->_redis) {
-			if ( !$this->_redis->select($db) ) {
+			if ( !$this->_redis->select((int)$db) ) {
 				return FALSE;
 			}
 			return $this->_redis->flushDb();
@@ -102,7 +102,7 @@ class Redis extends BaseClass implements BaseInterface
 
 		foreach ($this->getInfo() as $key => $val) {
 			if (preg_match('/db[0-9]+/', $key)) {
-				$databases[] = str_replace('db', '', $key);
+				$databases[] = (int)str_replace('db', '', $key);
 			}
 		}
 		return ($databases);
